refactor: split onboard prompt path
